feat: readonly rubric
